WIP moving validation methods to IMPValidator and utility methods into new class IMFCompositionPlaylistUtils
